refactor(server): extrair app factory para src/app.js e limpar server.js.
