+++ .env.example
@@ -1,9 +1,9 @@
 # Token da API do Tiny ERP
-TINY_API_TOKEN=939fa2d0e901ea51dce66df0858e876c67103e240d756ed48c4c3cc34e42ef30
+TINY_API_TOKEN=seu_token_aqui
 
 # Chave de autenticação para o FastZap acessar as rotas /bot
 # Gere com: openssl rand -hex 32
-BOT_API_KEY=e2e7deb29821b601
+BOT_API_KEY=mfm_gere_com_openssl_rand_hex_32
 
 # Porta local de desenvolvimento (em produção o Fly.io injeta PORT=8080)
 PORT=3003
